requirements.txt
streamlit
opencv-python
numpy
scipy
matplotlib
moviepy
yt-dlp
librosa
tqdm
scikit-image
torch
torchvision
Pillow
